+++ src/layer/RLayerVectorTile.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import {VectorTile as LayerVectorTile} from 'ol/layer';
 import {VectorTile as SourceVectorTile} from 'ol/source';
-import type {Options as OLOptions} from 'ol/source/VectorTile';
-import FeatureFormat from 'ol/format/Feature';
+import type {Options} from 'ol/source/VectorTile';
+import FeatureFormat, {FeatureToFeatureClass} from 'ol/format/Feature';
+import {FeatureLike} from 'ol/Feature';
+import RenderFeature from 'ol/render/Feature';
 
-import {OLFeatureClass, RContext, RContextType} from '../context';
+import {RContext, RContextType} from '../context';
 import {default as RLayer, RLayerProps} from './RLayer';
 import {default as RFeature, RFeatureUIEvent} from '../RFeature';
 import {OLEvent, RlayersBase} from '../REvent';
@@ -12,15 +14,10 @@ import {FeatureHandlers, featureHandlersSymbol} from './RLayerBaseVector';
 import RStyle, {RStyleLike} from '../style/RStyle';
 import debug from '../debug';
 
-type IfAny<T, Y, N> = 0 extends 1 & T ? Y : N;
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-type Options = IfAny<OLOptions<OLFeatureClass>, OLOptions, OLOptions<OLFeatureClass>>;
-
 /**
  * @propsfor RLayerVectorTile
  */
-export interface RLayerVectorTileProps extends RLayerProps {
+export interface RLayerVectorTileProps<F extends FeatureLike = RenderFeature> extends RLayerProps {
     /** URL for the tiles, normal {x}{y}{z} convention applies */
     url: string;
     /**
@@ -29,8 +26,12 @@ export interface RLayerVectorTileProps extends RLayerProps {
      * ie switching from a static OpenLayers style to a RefObject is not supported
      */
     style?: RStyleLike;
-    /** Vector tile format */
-    format: FeatureFormat;
+    /**
+     * Vector tile format
+     *
+     * this property currently does not support dynamic updates
+     */
+    format: FeatureFormat<FeatureToFeatureClass<F>>;
     /**
      * Width of the frame around the viewport that shall be rendered too
      * so that the symbols, whose center is outside of the viewport,
@@ -39,46 +40,60 @@ export interface RLayerVectorTileProps extends RLayerProps {
      * this property currently does not support dynamic updates
      */
     renderBuffer?: number;
-    /* vector tile specific source Options */
-    extent?: Options['extent'];
-    overlaps?: Options['overlaps'];
-    state?: Options['state'];
-    tileClass?: Options['tileClass'];
-    tileSize?: Options['tileSize'];
-    tileGrid?: Options['tileGrid'];
-    tileLoadFunction?: Options['tileLoadFunction'];
-    tileUrlFunction?: Options['tileUrlFunction'];
-    transition?: Options['transition'];
-    zDirection?: Options['zDirection'];
+    /* vector tile specific source Options<F> */
+    extent?: Options<F>['extent'];
+    overlaps?: Options<F>['overlaps'];
+    state?: Options<F>['state'];
+    tileClass?: Options<F>['tileClass'];
+    tileSize?: Options<F>['tileSize'];
+    tileGrid?: Options<F>['tileGrid'];
+    tileLoadFunction?: Options<F>['tileLoadFunction'];
+    tileUrlFunction?: Options<F>['tileUrlFunction'];
+    transition?: Options<F>['transition'];
+    zDirection?: Options<F>['zDirection'];
     /** onClick handler for loaded features */
-    onClick?: (this: RLayerVectorTile, e: RFeatureUIEvent) => boolean | void;
+    onClick?: (this: RLayerVectorTile<F>, e: RFeatureUIEvent<F>) => boolean | void;
     /** onPointerMove handler for loaded features */
-    onPointerMove?: (this: RLayerVectorTile, e: RFeatureUIEvent) => boolean | void;
+    onPointerMove?: (this: RLayerVectorTile<F>, e: RFeatureUIEvent<F>) => boolean | void;
     /** onPointerEnter handler for loaded features */
-    onPointerEnter?: (this: RLayerVectorTile, e: RFeatureUIEvent) => boolean | void;
+    onPointerEnter?: (this: RLayerVectorTile<F>, e: RFeatureUIEvent<F>) => boolean | void;
     /** onPointerLeave handler for loaded features */
-    onPointerLeave?: (this: RLayerVectorTile, e: RFeatureUIEvent) => boolean | void;
+    onPointerLeave?: (this: RLayerVectorTile<F>, e: RFeatureUIEvent<F>) => boolean | void;
 }
 
 /**
  * A vector tile layer
  *
- * Supports loading of features from vector tile servers
+ * Supports loading of features from vector tile servers.
  *
- * Only the handlers can be dynamically modified
+ * Only the handlers can be dynamically modified.
  *
- * Requires an `RMap` context
+ * Requires an `RMap` context.
  *
  * It does not provide a vector layer context for JSX-declared `RFeature`s
- * and it is not compatible with RLayerVector
+ * and it is not compatible with RLayerVector.
+ *
+ * Since 3.0, RLayerVectorTile uses OpenLayers light-weight RenderFeatures by default.
+ *
+ * @example
+ * <RLayerVectorTile url={url} format={new MVT()} />
+ *
+ * If you need advanced feature functions, you can switch to using full blown OpenLayers features
+ * by using a `format` parser that returns Features and, if using TSX, specifying the right
+ * generic argument.
+ *
+ * @example
+ * <RLayerVectorTile<Feature> url={url} format={new MVT({featureClass: Feature})} />
  */
-export default class RLayerVectorTile extends RLayer<RLayerVectorTileProps> {
-    ol: LayerVectorTile;
-    source: SourceVectorTile;
+export default class RLayerVectorTile<F extends FeatureLike = RenderFeature> extends RLayer<
+    RLayerVectorTileProps<F>
+> {
+    ol: LayerVectorTile<F>;
+    source: SourceVectorTile<F>;
 
-    constructor(props: Readonly<RLayerVectorTileProps>, context?: React.Context<RContextType>) {
+    constructor(props: Readonly<RLayerVectorTileProps<F>>, context?: React.Context<RContextType>) {
         super(props, context);
-        this.source = new SourceVectorTile({
+        this.source = new SourceVectorTile<F>({
             url: this.props.url,
             format: this.props.format,
             projection: this.props.projection,
@@ -119,7 +134,7 @@ export default class RLayerVectorTile extends RLayer<RLayerVectorTileProps> {
         featureHandlers[ev]--;
     }
 
-    protected refresh(prevProps?: RLayerVectorTileProps): void {
+    protected refresh(prevProps?: RLayerVectorTileProps<F>): void {
         super.refresh(prevProps);
         const handlers = Object.keys(this.props)
             .filter((ev) => ev.startsWith('on'))