+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [3.0.0]
+
+-   Do not support multiple OpenLayers versions, link each `rlayers` to one OpenLayers minor version
+-   Support light-weight `RenderFeature`s, these are opt-out for vector tile layers and opt-in for vector layers
+-   Use the same level of typing as OpenLayers
+-   For OpenLayers 9.2.2
+
 ### [2.3.2] 2024-05-22
 
 -   Fix [#252](https://github.com/mmomtchev/rlayers/issues/252), stop event propagation when the event handler returns `false`